[mod] do not split walk function
This implied to change the walker's API by changing the special case
on_leaf to before_leaf and after_leaf.

This wouldn't be a big deal if there was no leakage of "stop" handling
everywhere. Furthermore you must never forget to call your parent
"before" and "after" functions.

It's not pretty but the only way for now since the "visitor"'s and the
"property"'s code are mixed.