[fix] fix comparing strings with None values
A test testing bad indentation using python3 led to a bad fix by adding
an UnexpectedIndentation exception. In fact, the root cause was
a discrepency between python2 and python3 for comparing None values to
string. The added exception and try catch was thus hiding this.

This commit fixes the problem correctly by handling None values
similarly in python2 and python3. It highlights also where does the None
value comes from.